requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.10.8
charset-normalizer==2.0.7
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
debugpy==1.5.0
decorator==5.1.0
defusedxml==0.7.1
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.9.0
django-elasticsearch-dsl==7.2.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
drf-yasg==1.20.0
elasticsearch==7.15.0
elasticsearch-dsl==7.4.0
entrypoints==0.3
idna==3.3
inflection==0.5.1
ipykernel==6.4.1
ipython==7.28.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==3.0.3
jupyter-client==7.0.6
jupyter-core==4.8.1
MarkupSafe==2.0.1
matplotlib-inline==0.1.3
mysqlclient==2.0.3
nest-asyncio==1.5.1
numpy==1.21.2
packaging==21.2
pandas==1.3.3
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.2
prompt-toolkit==3.0.20
Pygments==2.10.0
PyJWT==1.7.1
pyparsing==2.4.7
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.1
pywin32==302
pyzmq==22.3.0
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.0
uritemplate==4.1.1
urllib3==1.26.7
waitress==2.0.0
wcwidth==0.2.5
whitenoise==5.3.0
xmltodict==0.12.0
